Rename summarization TrendWeeklyRate to TrendWeeklyElaboration
