+++ src/datasets/summarizations/trend-weekly-elaboration.summarization.ts
